Bump version odlparent->6.0.0,mdsal->5.0.2
This adopts new version of MRI projects and adjusts for the changes
therein. Most notably DPs/DPNs are identified by Uint64, not
BigInteger.

+++ interfacemanager/interfacemanager-impl/src/main/java/org/opendaylight/genius/interfacemanager/renderer/ovs/confighelpers/OvsInterfaceConfigAddHelper.java
@@ -15,7 +15,6 @@ import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.MoreExecutors;
-import java.math.BigInteger;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -58,6 +57,7 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.group.types.rev131018.group
 import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.NodeConnectorId;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.params.xml.ns.yang.ovsdb.rev150105.OvsdbBridgeAugmentation;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -143,7 +143,7 @@ public final class OvsInterfaceConfigAddHelper {
             return;
         }
 
-        BigInteger dpId = parentRefs.getDatapathNodeIdentifier();
+        Uint64 dpId = parentRefs.getDatapathNodeIdentifier();
         if (dpId == null) {
             LOG.warn("dpid for interface: {} Not Found. No DPID provided. " + "Creation of OF-Port not supported.",
                     interfaceNew.getName());
@@ -215,7 +215,7 @@ public final class OvsInterfaceConfigAddHelper {
     }
 
     private long createLogicalTunnelSelectGroup(TypedWriteTransaction<Configuration> tx,
-        BigInteger srcDpnId, String interfaceName, int lportTag) {
+            Uint64 srcDpnId, String interfaceName, int lportTag) {
         long groupId = IfmUtil.getLogicalTunnelSelectGroupId(lportTag);
         Group group = MDSALUtil.buildGroup(groupId, interfaceName, GroupTypes.GroupSelect,
                                            MDSALUtil.buildBucketLists(Collections.emptyList()));
@@ -243,7 +243,7 @@ public final class OvsInterfaceConfigAddHelper {
     }
 
     private void setupTunnelFlowsAndMonitoring(Interface interfaceNew, TypedWriteTransaction<Configuration> confTx,
-                                               IfTunnel ifTunnel, BigInteger dpId,
+                                               IfTunnel ifTunnel, Uint64 dpId,
                                                List<ListenableFuture<Void>> futures) {
         // if TEP is already configured on switch, start LLDP monitoring and
         // program tunnel ingress flow
